заметки мелкие правки + добавлена проверка на авторизацию

+++ task/routers.py
@@ -4,42 +4,50 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from database.database import get_async_session
 from database.models import User, Task
 from auth.routers import current_user
-from task.schemas import GetTaskSchema, AddTaskSchema, PatchTaskSchema
+from task.schemas import AddTaskSchema, PatchTaskSchema
 
 
 router = APIRouter()
 
 
 @router.get('/get/{pk}')
-async def get_current_task(pk: int, session: AsyncSession = Depends(get_async_session)):
-    await session.get(Task, pk)
-    return await session.get(Task, pk)
+async def get_current_task(pk: int, session: AsyncSession = Depends(get_async_session), user: User = Depends(current_user)):
+    result = await session.get(Task, pk)
+    if result.owner == user.email:
+        return result
+    else:
+        return {"msg": "you don’t have access rights"}
 
 
 @router.get('/get_all')
 async def get_all_tasks(session: AsyncSession = Depends(get_async_session), user: User = Depends(current_user)):
     query = select(Task).where(Task.owner == user.email)
-    res = await session.execute(query)
-    result = [GetTaskSchema.model_validate(row[0], from_attributes=True) for row in res.all()]
-    return result
+    result = await session.scalars(query)
+    return result.all()
 
 
 @router.post('/add')
-async def add_new_task(request: AddTaskSchema, session: AsyncSession = Depends(get_async_session)):
-    session.add(Task(**dict(request)))
+async def add_new_task(request: AddTaskSchema, session: AsyncSession = Depends(get_async_session),
+                       user: User = Depends(current_user)):
+    session.add(Task())
     await session.commit()
 
 
 @router.patch('/patch/{pk}')
-async def patch_task(pk: int, task: PatchTaskSchema, session: AsyncSession = Depends(get_async_session)):
+async def patch_task(pk: int, task: PatchTaskSchema, session: AsyncSession = Depends(get_async_session),
+                     user: User = Depends(current_user)):
     old_data = await session.get(Task, pk)
-    old_data.title = task.title
-    old_data.data = task.data
-    await session.commit()
+    if old_data.owner == user.email:
+        for key, value in task:
+            setattr(old_data, key, value)
+        await session.commit()
+        return {"msg": "Запись успешно обновлена"}
 
 
 @router.delete('/del/{pk}')
-async def delete_task(pk: int, session: AsyncSession = Depends(get_async_session)):
+async def delete_task(pk: int, session: AsyncSession = Depends(get_async_session), user: User = Depends(current_user)):
     task = await session.get(Task, pk)
-    await session.delete(task)
-    await session.commit()
+    if task.owner == user.email:
+        await session.delete(task)
+        await session.commit()
+        return {"msg": f"Запись {task.title} удалена"}
